requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cssselect2==0.8.0
distro==1.9.0
EbookLib==0.18
et_xmlfile==2.0.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
mammoth==1.9.0
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.6.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.29.0
networkx==3.4.2
numpy==2.2.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
pypdfium2==4.30.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.43.0
surya-ocr==0.13.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
unoconv==0.9.0
urllib3==2.3.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
XlsxWriter==3.2.2
zopfli==0.2.3.post1
